+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.2.1
+ * Fixed 5.1 bugs
+ * NOTE: With 5.1 petids were changed with guids so you're random summon list has to be configured from start
+
 ## 2.2
  * Tooltips of wild battle pets now have information if this companion is missing in your collection
  * While in pet battles, chat messages will inform you if the enemy pet is missing from your collection or if it has a higher quality than the same companion of your collection